+++ src/deployment/keycloak/index.ts
@@ -1,2 +1,13 @@
 export { KeycloakHelper } from "./deployment.js";
-export type { KeycloakUserConfig, KeycloakGroupConfig } from "./types.js";
+export {
+  DEFAULT_GROUPS,
+  DEFAULT_KEYCLOAK_CONFIG,
+  DEFAULT_USERS,
+} from "./constants.js";
+export type {
+  KeycloakUserConfig,
+  KeycloakGroupConfig,
+  KeycloakLdapFederationConfig,
+  KeycloakLdapRealmOptions,
+  KeycloakClientConfig,
+} from "./types.js";